Tag v0.2.0

+++ extension-al-object-designer/src/ALObjectCollectorCache.ts
@@ -25,7 +25,9 @@ export class ALObjectCollectorCache implements ALObjectDesigner.ObjectCollectorC
         // remove old cache folder: as of v0.2.0
         let root = (workspace as any).workspaceFolders[0];
         let oldPath = path.join(root.uri.fsPath, '.alcache', path.sep);
-        await fs.remove(oldPath);
+        if (await utils.folderExists(oldPath)) {
+            await fs.remove(oldPath);
+        }        
         
         let cache = await this.getCacheInfo(filepath, tagName);
         let stat = await fs.stat(filepath);
